# docs/api.yaml
openapi: 3.0.4
info:
  title: Detective API
  version: v1
paths:
...
  /api/v1/users:
    post:
      tags:
        - Users
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/UserType'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
        '403':
          description: Нет доступа
        '409':
          description: Conflict
        '503':
          description: Service Unavailable
    delete:
      tags:
        - Users
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '403':
          description: Нет доступа
        '404':
          description: Not Found
        '503':
          description: Service Unavailable
    get:
      tags:
        - Users
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '403':
          description: Нет доступа
        '503':
          description: Service Unavailable
...
components:
  schemas:
...
    UserDto:
      type: object
      properties:
        username:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/UserTypeDto'
      additionalProperties: false
    UserType:
      enum:
        - Employee
        - Admin
        - SpecialUser
      type: string
    UserTypeDto:
      enum:
        - Employee
        - Admin
        - SpecialUser
      type: string
  securitySchemes:
    Bearer:
      type: http
      description: Please enter a valid token
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]